complete unsigned

+++ srcs/ft_printfu.c
@@ -6,7 +6,7 @@
 /*   By: yichan <yichan@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/30 20:02:24 by yichan            #+#    #+#             */
-/*   Updated: 2022/08/04 20:55:56 by yichan           ###   ########.fr       */
+/*   Updated: 2022/08/04 21:36:18 by yichan           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -52,14 +52,14 @@ static void	lenadjust(t_flag *fmt, size_t fulllen, long nbr,
 
 void	ft_printfu(t_flag *fmt)
 {
-	unsigned long			nbrlen;
+	size_t			nbrlen;
 	size_t			fulllen;
 	long			nbr;
 	unsigned long	absolute;
 
 	fmt->precisionstate = fmt->precision;
 	fulllen = 0;
-	nbr = (long)va_arg(fmt->args, int);
+	nbr = (long)va_arg(fmt->args, unsigned);
 	if (nbr == 0 && fmt->dot && !fmt->width && !fmt->precision)
 		return ;
 	if (nbr < 0)
